[android] - fix test app runtime permissions
